requirements.txt
walrus
python-dotenv